Replace the top-of-screen snow-eating w/ hour drop
Remove the old chevron-eating code for when a snow column fills, which
was a temporary and somewhat ugly hack (and not quite sufficient: I was
running it with qh at h_/2, and remove * an extra 4).

Instead, get the time and let the bottom row fall away for the first
fifteen seconds of every hour. Get the time in simulate(), since this
still intends to be a clock at some point.