Make the snow collide with things.
This took forever to debug...becuase I forgot the the angle of repose
check *also* makes it fall and needed to check obstructions, so it was
eating it unless I had the fall check off and a one-pixel gap. Argh.

Anyway the digital clock can now be queried for collision, and the
static snow simulator can accept a functor to query for collision, and
it's all plumbed together and working.

And there's a whole second stupid rendering path from the clock that
reads the pixels back and re-sets them one by one to verify said
collision call.

Also some asserts, because nothing *should* be asking for out-of-range
collisions, but it's good to catch that if so, rather than debug
exciting memory corruptions. C APIs, yay.