More parameter tuning
Increase dynamic particles dramatically. Tried to make this half the
screen's worth, so it could handle the drop-off without exhaustion,
which leads to then having to rescan the whole framebuffer every tick
until it has finished, but I think simply scanning the huge array for
active particles is now a problem. Needs to become dynamic, and
defragmentable, I guess.

Give up and drop the tickrate again back down to 10Hz, and increase the
drip chance to compensate.