Switch the particles over to vector
Since it's super-expensive to not be able to spawn dynamic particles,
switch to a vector that can grow. Since it's super-expensive to scan
over hundreds of thousands of inactive particles, defragment and compact
it.

Midway point. This could drop the whole circular buffer thing entirely,
and probably should, since the upside and downsize fight and don't
really need to; we can let vector deal with the expotential physical
size growth, and just grow linearly logically.