Pop all static sand on the hour
The dropout is a little dull with every particle activating at once; on
particular on the cheap LCD mounted to the Pi it basically halves the
brightness as they dither with black then fall away. Pop instead, which
causes them to spray in all directions, and is a one-shot simulate. This
may also redistribute sand to the left a little bit.

Replace DEBUG_DROPOUT hacks with a more general fastclock flag that
effectively shifts the digits to MM:SS. It doesn't widen the dropout
gap, so it's only for a second, but it works well enough for testing.