Stop simming particles while all inactive
It's still surprisingly expensive, it seems, and since we're counting
the active particles, shut down even that when it reaches zero.

Since defragmenting is a bunch of moves but *not* a memory reallocation
at least, drop the defragment threshold more aggressively too. Sometimes
one particle will just manage to stay alive a long time.