Staticize particles that get stuck
After some frustrating debugging, there are some particles which are
getting stuck such that their proposed move is blocked in both the X and
Y direction and gravity isn't being enough to power them through (I
think this can happen if a segment splats on top of them, but beyond
that it may be "be sloppy with floating point/integer conversions at
your peril").

If a particle is unable to make progress, convert it to static. If it's
legtimiately free, it'll re-pop anyway, which will cause some weird
inertia, but better than it keeping the sim alive.

Rewrites the liveness condition to be easier to slap debugger points on,
and trusts the compiler to realize it can lazy-evaluate things.