SDL 2 support (without removing SDL 1)
Second binary, which is actually partially because I plan to stick a
menu on this thing for HyperPixel touchscreen purposes.

Slightly ugly SDLVERSION hacks, but mostly the two are compatible. The
big changes are contained to the separate main binary, needing to do
different graphics setup.

render() must now use the provided, not remembered, framebuffer, and
report if it wants to update the screen, which the main binary takes
care of (flip for SDL1, more complicated dance for SDL2).

Some changes around palette handling. Also surface creation---SDL2 is
more pedantic about complaining if you hand it a pixel format bitfield
at the same time as saying "paletted, please".

VSCode noise.

Flag which hacks are working or misbehaving weirdly under SDL2.

Yes, this all needs a pile of refactoring, but as always time is short.