Days since it was memory corruption: 0
The temporary surface was still being read in a bunch of places for
dimensions, which led to things like all the snow thinking it was out of
bounds. I have no idea how it caused a double-free inside vector once,
but that's undefined behaviour for you.

Stop hacking it about and instead deny the c'tors access to the surface,
giving them only the dimensions (and, for two awkward cases, pixel
formats that will *probably* still be the same later).

Now everything works with SDL2, and I don't *think* I've regressed 1.