Framework for menu support for SDL2 version
Currently the "menu" is a blank screen, but clicking switches to and
from it, pausing the background hack, and using an event-driven loop
rather than a fixed-tickrate one.

Also bring in libconfuse to the 2 version so the menu can actually
persist settings. This finally also means hacks can be changed at
startup, if you hand-craft a config for now.

Have hacks indicate they want to repaint the screen in advance, as
opposed to after calling render, to avoid the setup costs in SDL 2.

Tell VSCode to assume SDLVERSION 2.